templatize DiagnosticsInterface
